Cambiado un tipo mal puesto

+++ label_list.h
@@ -14,7 +14,7 @@ struct LABEL_LIST_NODE {
 
 void free_label_list(struct LABEL_LIST *list);
 
-int add_to_list(struct LABEL_LIST *list, char *label, unsigned long num_ins, int allow_duplicates);
+int add_to_list(struct LABEL_LIST *list, char *label, unsigned int num_ins, int allow_duplicates);
 
 void debug_print_label_list(struct LABEL_LIST *list);
